Update Index.rst
Adding "only" to the phrase on line 22 seems to add meaning to phrase after it. I may be wrong about the meaning or the purpose. Made some changes in phrase order. Seeking to focus on the active sense and placing attention on word order for fluid thinking (i.e., placing the adverb 'artfully' closer to its verb, 'combine').